Split rest server out of main.go
This is just preparing for adding the grpc auth proxy.
